ai-photo-editor/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.4.0
gunicorn==22.0.0
